Fix character level up to 50
The character does not receive life, mana, and skill points when leveling up to 50 because the NextPlrLevel function is skipped upon reached the needed experience to level up to 50. This change allows the player to receive the benefits of leveling up, but still prevents the player from leveling up to 51.